Testing & Error Resolving

+++ backend/controllers/RestaurantMenuController.php
@@ -2,17 +2,15 @@
 
 namespace backend\controllers;
 
-use Yii;
-use common\models\RestaurantMenu;
-use common\models\RestaurantMenuSearch;
-use yii\web\Controller;
-use yii\web\NotFoundHttpException;
-use yii\filters\VerbFilter;
 use backend\components\AdminCoreController;
 use common\components\Common;
 use common\models\MenuCategories;
-use yii\helpers\ArrayHelper;
-
+use common\models\RestaurantMenu;
+use common\models\RestaurantMenuSearch;
+use Yii;
+use yii\web\BadRequestHttpException;
+use yii\web\Controller;
+use yii\web\NotFoundHttpException;
 
 /**
  * RestaurantMenuController implements the CRUD actions for RestaurantMenu model.
@@ -24,14 +22,14 @@ class RestaurantMenuController extends AdminCoreController
      */
     /*public function behaviors()
     {
-        return [
-            'verbs' => [
-                'class' => VerbFilter::className(),
-                'actions' => [
-                    'delete' => ['POST'],
-                ],
-            ],
-        ];
+    return [
+    'verbs' => [
+    'class' => VerbFilter::className(),
+    'actions' => [
+    'delete' => ['POST'],
+    ],
+    ],
+    ];
     }*/
 
     /**
@@ -41,20 +39,25 @@ class RestaurantMenuController extends AdminCoreController
     public function actionIndex()
     {
         $snRestaurantId = ($_GET['rid'] > 0) ? $_GET['rid'] : 0;
-        $snRestaurantName = Common::get_name_by_id($snRestaurantId,$flag = "Restaurants");
-        $searchModel = new RestaurantMenuSearch();
-        $dataProvider = $searchModel->backendSearch(Yii::$app->request->queryParams);
-        $MenuCategories = MenuCategories::MenuCategoriesDropdown();
-    
-      //  p($MenuCategories);
+        $user = Common::get_user_role(Yii::$app->user->id, $flag = "1");
+        if ((($user->role_id == Yii::$app->params['userroles']['manager']) && ($user->restaurant_id == $snRestaurantId)) || (($user->role_id == Yii::$app->params['userroles']['admin']) || ($user->role_id == Yii::$app->params['userroles']['super_admin']))) {
+            $snRestaurantName = Common::get_name_by_id($snRestaurantId, $flag = "Restaurants");
+            $searchModel = new RestaurantMenuSearch();
+            $dataProvider = $searchModel->backendSearch(Yii::$app->request->queryParams);
+            $MenuCategories = MenuCategories::MenuCategoriesDropdown();
 
+            //  p($MenuCategories);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-            'snRestaurantName' => $snRestaurantName,
-            'MenuCategories' => $MenuCategories,
-        ]);
+            return $this->render('index', [
+                'searchModel' => $searchModel,
+                'dataProvider' => $dataProvider,
+                'snRestaurantName' => $snRestaurantName,
+                'MenuCategories' => $MenuCategories,
+            ]);
+        } else {
+            throw new BadRequestHttpException('The requested page does not exist.');
+
+        }
     }
 
     /**
@@ -80,31 +83,36 @@ class RestaurantMenuController extends AdminCoreController
     {
         $model = new RestaurantMenu();
         $snRestaurantId = ($_GET['rid'] > 0) ? $_GET['rid'] : 0;
-        $snRestaurantName = Common::get_name_by_id($snRestaurantId,$flag = "Restaurants");
-        $model->restaurant_id = $_GET['rid'];
-        $MenuCategoriesDropdown = MenuCategories::MenuCategoriesDropdown();
-        
-        if ($model->load(Yii::$app->request->post())) {
-             $file = \yii\web\UploadedFile::getInstance($model, 'photo');
-            if (!empty($file)){   
-                $file_name = $file->basename."_".uniqid().".".$file->extension;
-                $model->photo = $file_name;
-                if($model->validate()){
-                    $model->save();
-                    $file->saveAs( Yii::getAlias('@root') .'/frontend/web/uploads/' . $file_name);
+        $user = Common::get_user_role(Yii::$app->user->id, $flag = "1");
+        if ((($user->role_id == Yii::$app->params['userroles']['manager']) && ($user->restaurant_id == $snRestaurantId)) || (($user->role_id == Yii::$app->params['userroles']['admin']) || ($user->role_id == Yii::$app->params['userroles']['super_admin']))) {
+            $snRestaurantName = Common::get_name_by_id($snRestaurantId, $flag = "Restaurants");
+            $model->restaurant_id = $_GET['rid'];
+            $MenuCategoriesDropdown = MenuCategories::MenuCategoriesDropdown();
 
+            if ($model->load(Yii::$app->request->post())) {
+                $file = \yii\web\UploadedFile::getInstance($model, 'photo');
+                if (!empty($file)) {
+                    $file_name = $file->basename . "_" . uniqid() . "." . $file->extension;
+                    $model->photo = $file_name;
+                    if ($model->validate()) {
+                        $model->save();
+                        $file->saveAs(Yii::getAlias('@root') . '/frontend/web/uploads/' . $file_name);
+
+                    }
+                    Yii::$app->session->setFlash('success', Yii::getAlias('@restaurant_menu_add_message'));
+                    return $this->redirect(['index', 'rid' => $model->restaurant_id]);
+                } else {
+                    return $this->render('create', ['model' => $model, 'snRestaurantName' => $snRestaurantName, 'MenuCategoriesDropdown' => $MenuCategoriesDropdown]);
                 }
-                Yii::$app->session->setFlash( 'success', Yii::getAlias( '@restaurant_menu_add_message' ) );
-                return $this->redirect(['index', 'rid' => $model->restaurant_id]);
-            }else{
-                return $this->render('create', ['model' => $model,'snRestaurantName' => $snRestaurantName,'MenuCategoriesDropdown'=>$MenuCategoriesDropdown]);
             }
+            return $this->render('create', [
+                'model' => $model,
+                'snRestaurantName' => $snRestaurantName,
+                'MenuCategoriesDropdown' => $MenuCategoriesDropdown,
+            ]);
+        } else {
+            throw new BadRequestHttpException('The requested page does not exist.');
         }
-        return $this->render('create', [
-            'model' => $model,
-            'snRestaurantName' => $snRestaurantName,
-            'MenuCategoriesDropdown'=>$MenuCategoriesDropdown,
-        ]);
     }
 
     /**
@@ -117,40 +125,44 @@ class RestaurantMenuController extends AdminCoreController
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
-         $snRestaurantId = ($_GET['rid'] > 0) ? $_GET['rid'] : 0;
-        $old_image = $model->photo;
-        $snRestaurantName = Common::get_name_by_id($snRestaurantId,$flag = "Restaurants");
-         $MenuCategoriesDropdown = MenuCategories::MenuCategoriesDropdown();
-        if ($model->load(Yii::$app->request->post())) {
-             $file = \yii\web\UploadedFile::getInstance($model, 'photo');
-      
-           if (!empty($file)){
-                 $delete = $model->oldAttributes['photo'];
-                 $file_name = $file->basename."_".uniqid().".".$file->extension;
+        $snRestaurantId = ($_GET['rid'] > 0) ? $_GET['rid'] : 0;
+        $user = Common::get_user_role(Yii::$app->user->id, $flag = "1");
+        if ((($user->role_id == Yii::$app->params['userroles']['manager']) && ($user->restaurant_id == $snRestaurantId)) || (($user->role_id == Yii::$app->params['userroles']['admin']) || ($user->role_id == Yii::$app->params['userroles']['super_admin']))) {
+            $old_image = $model->photo;
+            $snRestaurantName = Common::get_name_by_id($snRestaurantId, $flag = "Restaurants");
+            $MenuCategoriesDropdown = MenuCategories::MenuCategoriesDropdown();
+            if ($model->load(Yii::$app->request->post())) {
+                $file = \yii\web\UploadedFile::getInstance($model, 'photo');
 
-                 $model->photo = $file_name; 
-            if(!empty($old_image) && file_exists(Yii::getAlias('@root') .'/frontend/web/uploads/'.$old_image)){
-                 unlink(Yii::getAlias('@root') .'/frontend/web/uploads/'.$old_image);
-             }
-                 $file->saveAs( Yii::getAlias('@root') .'/frontend/web/uploads/' . $file_name,false);
-                 $model->photo = $file_name;
-                 $model->save();
+                if (!empty($file)) {
+                    $delete = $model->oldAttributes['photo'];
+                    $file_name = $file->basename . "_" . uniqid() . "." . $file->extension;
+
+                    $model->photo = $file_name;
+                    if (!empty($old_image) && file_exists(Yii::getAlias('@root') . '/frontend/web/uploads/' . $old_image)) {
+                        unlink(Yii::getAlias('@root') . '/frontend/web/uploads/' . $old_image);
+                    }
+                    $file->saveAs(Yii::getAlias('@root') . '/frontend/web/uploads/' . $file_name, false);
+                    $model->photo = $file_name;
+                    $model->save();
+                } else {
+                    $model->photo = $old_image;
+                    $model->save(false);
+                }
+                Yii::$app->session->setFlash('success', Yii::getAlias('@restaurant_menu_update_message'));
+                return $this->redirect(['index', 'rid' => $model->restaurant_id]);
+            } else {
+                return $this->render('update', ['model' => $model, 'snRestaurantName' => $snRestaurantName, 'MenuCategoriesDropdown' => $MenuCategoriesDropdown]);
             }
-            else{
-                $model->photo  = $old_image;
-                $model->save(false);
-            }
-            Yii::$app->session->setFlash( 'success', Yii::getAlias( '@restaurant_menu_update_message' ) );
-            return $this->redirect(['index','rid' => $model->restaurant_id]);
-        }else{
-            return $this->render('update', ['model' => $model,'snRestaurantName' => $snRestaurantName,'MenuCategoriesDropdown'=>$MenuCategoriesDropdown]);
+
+            return $this->render('update', [
+                'model' => $model,
+                'snRestaurantName' => $snRestaurantName,
+                'MenuCategoriesDropdown' => $MenuCategoriesDropdown,
+            ]);
+        } else {
+            throw new BadRequestHttpException('The requested page does not exist.');
         }
-
-        return $this->render('update', [
-            'model' => $model,
-            'snRestaurantName' => $snRestaurantName,
-            'MenuCategoriesDropdown'=>$MenuCategoriesDropdown,
-        ]);
     }
 
     /**
@@ -162,15 +174,22 @@ class RestaurantMenuController extends AdminCoreController
      */
     public function actionDelete($id)
     {
-       $model = $this->findModel($id);
-       $snDeleteStatus = $model->delete();
-        if ( !empty( $snDeleteStatus ) && $snDeleteStatus=='1' ) {
-            if(file_exists(Yii::getAlias('@root') . '/frontend/web/uploads/'. $model->photo))
-            unlink(Yii::getAlias('@root') . '/frontend/web/uploads/'. $model->photo);
-            $model->delete(); 
-            Yii::$app->session->setFlash( 'success', Yii::getAlias( '@restaurant_menu_delete_message' ) );
+        $model = $this->findModel($id);
+        $user = Common::get_user_role(Yii::$app->user->id, $flag = "1");
+        if ((($user->role_id == Yii::$app->params['userroles']['manager']) && ($user->restaurant_id == $model->restaurant_id)) || (($user->role_id == Yii::$app->params['userroles']['admin']) || ($user->role_id == Yii::$app->params['userroles']['super_admin']))) {
+            $snDeleteStatus = $model->delete();
+            if (!empty($snDeleteStatus) && $snDeleteStatus == '1') {
+                if (file_exists(Yii::getAlias('@root') . '/frontend/web/uploads/' . $model->photo)) {
+                    unlink(Yii::getAlias('@root') . '/frontend/web/uploads/' . $model->photo);
+                }
+
+                $model->delete();
+                Yii::$app->session->setFlash('success', Yii::getAlias('@restaurant_menu_delete_message'));
+            }
+            return $this->redirect(['index', 'rid' => $model->restaurant_id]);
+        } else {
+            throw new BadRequestHttpException('The requested page does not exist.');
         }
-        return $this->redirect(['index','rid' => $model->restaurant_id]);
     }
 
     /**
